+++ airflow/models/dag.py
@@ -90,6 +90,7 @@ from airflow.typing_compat import Literal
 from airflow.utils import timezone
 from airflow.utils.dag_cycle_tester import check_cycle
 from airflow.utils.dates import cron_presets, date_range as utils_date_range
+from airflow.utils.decorators import fixup_decorator_warning_stack
 from airflow.utils.file import correct_maybe_zipped
 from airflow.utils.helpers import at_most_one, exactly_one, validate_key
 from airflow.utils.log.logging_mixin import LoggingMixin
@@ -3543,6 +3544,8 @@ def dag(
             # Return dag object such that it's accessible in Globals.
             return dag_obj
 
+        # Ensure that warnings from inside DAG() are emitted from the caller, not here
+        fixup_decorator_warning_stack(factory)
         return factory
 
     return wrapper
